+++ docs/additional_info/changelog.md
@@ -2,6 +2,36 @@
 ---
 # Changelog
 
+## 16.3.0 (17-Jul-2026)
+
+Added support for the new [Audit Logs endpoint](https://developers.lokalise.com/reference/list-audit-logs).
+
+Audit Logs are available through the new `LokaliseApiV1` client:
+
+```ts
+const lokaliseApi = new LokaliseApiV1({
+	apiKey: process.env.API_KEY,
+});
+```
+
+Example:
+
+```ts
+const auditLogs = await lokaliseApi.auditLogs().list({
+	// Optional parameters:
+	limit: 100,
+	cursor: "some_cursor",
+	event_type: "project.deleted",
+});
+
+console.log(auditLogs.items);
+console.log(auditLogs.items[0].metadata.event_code); // "project.deleted"
+console.log(auditLogs.next_cursor);
+console.log(auditLogs.has_more);
+```
+
+More endpoints will gradually become available through `LokaliseApiV1`.
+
 ## 16.2.0 (15-Jul-2026)
 
 Added support for upload from FSS: